Align matrix poll code in main.commented to that of main.c
The code in the commented main file probed the matrix row by row while
the current main code probes the matrix column by column. This commit
changes the commented code to probe in the same way the main code does.

This commit fixes also a small typo in the description of the return
value of `probe_column` (it returns `0b00000000` for non-pressed keys,
not `0b00000001`).